Switched to SMT-LIB Implementation as main
